+++ app/Http/Controllers/ProductController.php
@@ -19,12 +19,14 @@ class ProductController extends Controller
 
     public function index(Request $request)
     {
+        $this->authorize('viewAny', Product::class);
         $items = $this->productService->all($request);
         return view('admin.product.index', compact('items'));
     }
 
     public function create()
     {
+        $this->authorize('create', Product::class);
         $categories = Category::all();
         return view('admin.product.create', compact('categories'));
     }
@@ -49,6 +51,7 @@ class ProductController extends Controller
 
     public function edit($id)
     {
+        $this->authorize('update', Product::class);
         $items = $this->productService->find($id);
         $categories = Category::all();
         $product = $this->productService->find($id);
@@ -71,6 +74,7 @@ class ProductController extends Controller
 
     public function destroy($id)
     {
+        $this->authorize('delete', Product::class);
         $items = $this->productService->delete($id);
         try {
             toast('Xóa Tạm Thời Sản Phẩm Thành Công!','success','top-right');
@@ -102,6 +106,10 @@ class ProductController extends Controller
     }
     public function deleteforever($id)
     {
+
+        $this->authorize('deleteforever', Product::class);
+        $items = $this->productService->deleteforever($id);
+
         try {
             $items = $this->productService->deleteforever($id);
             toast('Xóa Vĩnh Viễn Sản Phẩm Thành Công!','success','top-right');
@@ -111,6 +119,7 @@ class ProductController extends Controller
             toast('Xóa Vĩnh Viễn Sản Phẩm Không Thành Công!','danger','top-right');
             return redirect()->route('products.index');
         }
+
         return redirect()->route('products.index');
     }
 }
